requirements.txt
langchain
langchain-ollama
langchain-community
langgraph
chromadb
python-dotenv
pydantic
pydantic-settings
chainlit